pipeline

+++ scripts/s2orc/memes_pipeline.py
@@ -1,6 +1,7 @@
 
 import embedd_noun_chunks
 import cluster
+import os
 import reduce_dimensionality
 import create_meme_names
 from meme_score import meme_score
@@ -18,6 +19,8 @@ def main(   out_path_embedding:str,
             in_path:str='data/s2orc/processed_big.parquet',
             cit_path = r'C:/Users/ppaul/Documents/AI-strategies-papers-regulations-monitoring/data/s2orc/big_ai_dataset.parquet', 
             sentences_embedding: str = "all-MiniLM-L6-v2", 
+            json_input = r'C:/Users/ppaul/Documents/AI-strategies-papers-regulations-monitoring/data/s2orc/doi_to_authorship_big.json',
+            aff_output_path = 'data/s2orc/big_ai_dataset_with_affiliations.parquet',
             batch_size_embedd:int=32, 
             multiprocess_embedd:bool=False, 
             do_reduce:bool = True, 
@@ -28,43 +31,43 @@ def main(   out_path_embedding:str,
             gpu_cluster: bool = False, 
             min_clust_size: int = 5, 
             metric_cluster: str = 'euclidean',
-            conditioning_meme_score:str='is_big_tech',
             condition_list:list = ['PL'],
-            category:str = 'country',
-            json_input = r'C:/Users/ppaul/Documents/AI-strategies-papers-regulations-monitoring/data/s2orc/doi_to_authorship_big.json',
-            aff_output_path = 'data/s2orc/big_ai_dataset_with_affiliations.parquet'):
+            category:str = 'country'
+            ):
     #embedd noun_chunks
-    print('EMBEDDING NOUN CHUNKS')
-    chunk_to_embedding = embedd_noun_chunks.embedd_noun_chunks(in_path, sentences_embedding,batch_size_embedd,multiprocess_embedd)#return chunk to embeddinng
-    
-    #reduce dimensionality
-    if do_reduce:
-        print('REDUCING')
-        reduced = reduce_dimensionality.reducing(chunk_to_embedding, n_components_reduce, gpu_reduce)
+    if not os.path.exists(out_path_cluster):
+        print('EMBEDDING NOUN CHUNKS')
+        chunk_to_embedding = embedd_noun_chunks.embedd_noun_chunks(in_path, sentences_embedding,batch_size_embedd,multiprocess_embedd)#return chunk to embeddinng
+        
+        #reduce dimensionality
+        if do_reduce:
+            print('REDUCING')
+            reduced = reduce_dimensionality.reducing(chunk_to_embedding, n_components_reduce, gpu_reduce)
 
-    pd.DataFrame(chunk_to_embedding).to_parquet(out_path_embedding)#saving
+        pd.DataFrame(chunk_to_embedding).to_parquet(out_path_embedding)#saving
 
-    print('CLUSTERING')
-    df_cluster = cluster.clustering(reduced,
-         n_jobs_cluster,
-         cluster_selection_epsilons,
-         gpu_cluster,
-         min_clust_size,
-         metric_cluster)
+        print('CLUSTERING')
+        df_cluster = cluster.clustering(reduced,
+            n_jobs_cluster,
+            cluster_selection_epsilons,
+            gpu_cluster,
+            min_clust_size,
+            metric_cluster)
 
-
-    pd.DataFrame(df_cluster).to_parquet(out_path_cluster)#saving
+        pd.DataFrame(df_cluster).to_parquet(out_path_cluster)#saving
+    else:
+        df_cluster = pd.read_parquet(out_path_cluster)
 
     print('GETTING AFFILIATIONS')
     df_aff = affiliation_pipeline.affiliations(condition_list,category,cit_path,json_input,aff_output_path)
     
-    #dodać affiliacje!!!!!!
     print('PREPARING MEMES')
     df_cluster, chunk_to_meme = prepare_memes.preparing(df_cluster,in_path,df_aff)
     print('CREATING NAMES')
     meme_to_name = create_meme_names.names(chunk_to_meme,in_path)
+
     print('CALCULATING MEME SCORE')
-    df_meme_score = meme_score(df_cluster, conditioning_meme_score)
+    df_meme_score = meme_score(df_cluster)
 
     df_meme_score['meme_name'] = df_meme_score['meme_id'].map(meme_to_name["best_tfidf"])
